Add runtime list and runtime info api.

+++ src/main/java/romever/scan/oasisscan/vo/chain/NodeStatus.java
@@ -2,9 +2,11 @@ package romever.scan.oasisscan.vo.chain;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class NodeStatus {
     private boolean expiration_processed;
     private Object freeze_end_time;
-    private Long election_eligible_after;
+    private BigInteger election_eligible_after;
 }
